Ensure the order of the returned data points
The randomness was probably caused by some performance optimizations behind the scenes like multi-threaded access to the db. Unfortunately, the data needs to be ordered (like we add it) so we have to manually order it to be sure.
Also make sure data is streamed instead of pulling everything into memory first.