Stream data efficiently without limit
If you just use a normal ObjectResult, you'll run into MaxIAsyncEnumerableBufferLimit which (in our case) controls the maximum number of objects buffered before streaming. We exeed that by a lot so we need to stream it directly. The solution inspired by https://stackoverflow.com/a/26281258/10883465 works fine and is quite fast but I believe it could and should be optimized (see the code-comments).