Run the backup job inside the docker container
Previously it was an unprivileged cron job that would have used the psql installed on the system. Since we don't want to have to install psql on the rpi itself and instead use the one of the timescale image, we docker exec into it requiring the job to have root privileges.